refactor: rename root level utils to common dir
