+++ src/BreakpointsContext/utils/denormalizeBreakpointsState.util.ts/denormalizeBreakpointsState.util.ts
@@ -1,4 +1,4 @@
-import { set } from '../../../utils/set'
+import { set } from '../../../common/set'
 
 type DenormalizeBreakpointsStateUtil = (state: Rmb.NormalizedConfig) => Rmb.Breakpoints
 export const denormalizeBreakpointsState: DenormalizeBreakpointsStateUtil = state => {
