+++ src/initMatchBreakpoints/utils/normalizeMediaQueryDict/normalizeMediaQueryDict.util.ts
@@ -1,4 +1,4 @@
-import { isPlainObject } from '../../../utils/isPlainObject'
+import { isPlainObject } from '../../../common/isPlainObject'
 
 export const normalizeBreakpointsConfig = <T extends boolean | string>(
   breakpointsConfig: Rmb.Config<T>,
